+++ Configuration GUI/src/App.tsx
@@ -17,7 +17,7 @@
  * Steffen Lindner (steffen.lindner@uni-tuebingen.de)
  */
 
-import React, {useState} from 'react'
+import React, {useEffect, useState} from 'react'
 import Config from "./config"
 import {BrowserRouter as Router, Route, Routes} from "react-router-dom"
 import {Col, Container, Row} from "react-bootstrap"
@@ -33,6 +33,8 @@ import Settings from "./sites/Settings";
 import Offline from "./sites/Offline"
 import Tables from "./sites/Tables";
 import config from "./config";
+import {DefaultStream, DefaultStreamSettings, StreamSettings} from "./common/Interfaces";
+import {Stream} from "./common/Interfaces";
 
 const App = () => {
     const [error, set_error] = useState(false)
@@ -52,6 +54,43 @@ const App = () => {
     const Wrapper = styled.div`
 
     `
+
+    // Validates the stored streams and stream settings in the local storage
+    // Clears local storage if some streams/settings are not valid
+    // This may be needed if the UI got an update (new stream properties), but the local storage
+    // holds "old" streams/settings without the new property
+    const validateLocalStorage = () => {
+        const defaultStream = DefaultStream(1)
+        const defaultStreamSetting = DefaultStreamSettings(1, 5)
+
+        try {
+            let stored_streams: Stream[] = JSON.parse(localStorage.getItem("streams") ?? "[]")
+            let stored_settings: StreamSettings[] = JSON.parse(localStorage.getItem("streamSettings") ?? "[]")
+
+            if(!stored_streams.every(s => Object.keys(defaultStream).every(key => Object.keys(s).includes(key)))) {
+                alert("Incompatible stream description found. This may be due to an update. Resetting local storage.")
+                localStorage.clear()
+                window.location.reload()
+                return
+            }
+
+            if(!stored_settings.every(s => Object.keys(defaultStreamSetting).every(key => Object.keys(s).includes(key)))) {
+                alert("Incompatible stream description found. This may be due to an update. Resetting local storage.")
+                localStorage.clear()
+                window.location.reload()
+                return
+            }
+        }
+        catch {
+            alert("Error in reading local storage. Resetting local storage.")
+            localStorage.clear()
+            window.location.reload()
+        }
+    }
+
+    useEffect(() => {
+        validateLocalStorage()
+    }, [])
     return <>
         <Router basename={Config.BASE_PATH}>
             <Row>